requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
requests==2.31.0
